feat: integrate routes with ACL

+++ apps/backend/src/thoughts/reply.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
@@ -9,7 +10,10 @@ import {
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 
-import { AuthValidate, User } from '@app/utils';
+import { AuthValidate } from '@app/utils';
+import { AppAbility, User } from '@app/acl';
+
+import { subject } from '@casl/ability';
 import { CreateReplyDto, PaginationDto } from './dto';
 import { RepliesService } from './replies.service';
 import { createReply, deleteReply, findAllReplies } from './schemas';
@@ -23,20 +27,33 @@ export class ReplyController {
   @Post(':id/reply')
   @AuthValidate(createReply)
   async create(
-    @User('id') user: number,
+    @User('user.id') user: number,
+    @User('ability') ability: AppAbility,
     @Param('id') id: number,
     @Body() createReplyDto: CreateReplyDto
   ) {
+    if (!ability.can('create', 'replies')) {
+      throw new ForbiddenException(
+        'You are not authorized to perform this action'
+      );
+    }
     return this.repliesService.create(id, user, createReplyDto);
   }
 
   @Get(':id/reply')
   @AuthValidate(findAllReplies)
   async findAll(
-    @User('id') userId: number,
+    @User('user.id') userId: number,
+    @User('ability') ability: AppAbility,
     @Param('id') id: number,
     @Query() { latest, offset, limit }: PaginationDto
   ) {
+    if (!ability.can('read', 'replies')) {
+      throw new ForbiddenException(
+        'You are not authorized to perform this action'
+      );
+    }
+
     const { replies, count } = await this.repliesService.findAll(id, {
       latest,
       offset,
@@ -57,7 +74,19 @@ export class ReplyController {
 
   @Delete(':id/reply/:replyId')
   @AuthValidate(deleteReply)
-  async remove(@Param('id') id: number, @Param('replyId') replyId: number) {
+  async remove(
+    @User('ability') ability: AppAbility,
+    @Param('id') id: number,
+    @Param('replyId') replyId: number
+  ) {
+    const reply = await this.repliesService.findOne(replyId);
+
+    if (!ability.can('delete', subject('replies', reply))) {
+      throw new ForbiddenException(
+        'You are not authorized to perform this action'
+      );
+    }
+
     return this.repliesService.remove(id, replyId);
   }
 }
